+++ packages/streamline/src/components/input-date-picker/input-date-picker.tsx
@@ -2,14 +2,12 @@ import DateTimePicker, {
   DateTimePickerEvent,
 } from '@react-native-community/datetimepicker';
 import React, { useEffect } from 'react';
-import { Modal } from 'react-native';
 import { Pressable } from 'react-native-ama';
 
-import { Box } from '../../primitives/box/box';
-import { Text } from '../../primitives/text/text';
 import { DateFormat, formatDate } from '../../utils/date';
 import { isAndroid, isIOS } from '../../utils/platform';
 import { InputText } from '../inputs/input-text/input-text';
+import { ModalPickerIOS } from './modal-picker-ios';
 
 /* eslint-disable-next-line */
 export interface InputDatePickerProps {
@@ -114,90 +112,5 @@ export function InputDatePicker({
     </Pressable>
   );
 }
-
-interface ModalPickerIOS
-  extends Pick<
-    InputDatePickerProps,
-    'maximumDate' | 'minimumDate' | 'okLabelIOS'
-  > {
-  focused: boolean;
-  setFocused: (focused: boolean) => void;
-  innerOnChange?: () => void;
-  innerDate?: Date;
-}
-
-const ModalPickerIOS = ({
-  focused,
-  setFocused,
-  okLabelIOS = 'OK',
-  innerDate,
-  innerOnChange,
-  minimumDate,
-  maximumDate,
-}: ModalPickerIOS) => (
-  <Modal
-    animationType="fade"
-    transparent={true}
-    visible={focused}
-    onRequestClose={() => {
-      setFocused(false);
-    }}
-  >
-    <Pressable
-      accessibilityRole="button"
-      accessibilityLabel={okLabelIOS.toLocaleLowerCase()}
-      onPress={() => {
-        setFocused(false);
-      }}
-      style={{ flex: 1 }}
-    >
-      <Box
-        backgroundColor="BLACK"
-        opacity={0.2}
-        position="absolute"
-        bottom={0}
-        left={0}
-        right={0}
-        top={0}
-        flex={1}
-      />
-    </Pressable>
-    <Box
-      backgroundColor="GREY_100"
-      position="absolute"
-      bottom={0}
-      left={0}
-      right={0}
-    >
-      <Box
-        backgroundColor="PURE_WHITE_1000"
-        height={40}
-        width="100%"
-        alignItems="flex-end"
-        justifyContent="center"
-      >
-        <Text
-          testID="answer-done"
-          onPress={() => setFocused(false)}
-          color="BLUKO_500"
-          padding="xs"
-          variant="bodyBold"
-        >
-          {okLabelIOS}
-        </Text>
-      </Box>
-      <DateTimePicker
-        testID="date-picker-input"
-        value={innerDate || new Date()}
-        display="inline"
-        onChange={innerOnChange}
-        minimumDate={minimumDate}
-        maximumDate={maximumDate}
-        themeVariant="light"
-        style={{ height: 400 }}
-      />
-    </Box>
-  </Modal>
-);
 
 export default InputDatePicker;
